requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0

# AI / API
anthropic>=0.40.0

# Visualisierung
plotly>=5.18.0

# Webapp
streamlit>=1.32.0

# Daten laden
requests>=2.31.0
openpyxl>=3.1.0       # Excel-Dateien (BFS)

# Notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Utils
python-dotenv>=1.0.0  # .env für API Keys
tqdm>=4.66.0          # Fortschrittsbalken bei Batch-Scoring
